Move back temporary structures home.
